AI-3236: extract global-search path and shared models out of search.py
Addresses Martin's review (search.py was bloated, 1094 lines):
- search_models.py: shared models (SearchHit/SearchOutput/SearchSpec/PatternMatch),
  type aliases and constants used by both search paths.
- search_global.py: the server-side global-search textual path (_global_textual_search
  and helpers), depending only on search_models (no import cycle).
- search.py (1094 -> ~590 lines): tool entry point + legacy enumeration path; re-exports
  the public names via __all__ so existing importers are unaffected.

Pure refactor — no behavior change (full unit suite green; real search() verified against
canary-orion projects: not-indexed 198 still falls back, indexed 21 uses the fast path).
